service/requirements.txt
bidict==0.23.1
certifi==2025.1.31
charset-normalizer==3.4.1
h11==0.14.0
idna==3.10
python-engineio==4.11.2
python-socketio==5.12.1
requests==2.32.3
simple-websocket==1.1.0
urllib3==2.2.3
wsproto==1.2.0
